Refactor: Split manage_users function
Split the manage_users Supabase function into smaller files for improved maintainability.
