+++ supabase/functions/manage_users/types.ts
@@ -1,0 +1,18 @@
+
+export interface UserProfile {
+  id: string;
+  full_name: string | null;
+  email: string;
+  role: string;
+  plan: string;
+  first_name: string;
+  last_name: string;
+}
+
+export interface UpdateUserData {
+  id: string;
+  first_name?: string;
+  last_name?: string;
+  role?: 'pet_lover' | 'pet_caddie' | 'site_manager';
+  plan?: 'free' | 'premium' | 'teams' | 'personal';
+}
